Modularize MTP/Spec feature out of the three monoliths
This branch had piled a lot onto already-monolithic files. The MTP/Spec
feature has clean boundaries, so it can move out without entangling the
rest. Pure code organization — zero functional behavior changes.

New modules
-----------
- modules/spec_launch.py (441 lines): per-backend spec/MTP emission
  helpers (emit_spec_args, emit_reasoning_args, emit_kv_unify_args,
  emit_no_mmproj_arg, resolve_effective_parallel) plus the four
  spec-type whitelist frozensets. LaunchManager.build_cmd() now calls
  these helpers at the same positions as before, dropping ~320 lines
  of inline emission to ~10 lines of call sites.
- modules/spec_persistence.py (359 lines): cfg/app_settings mirroring
  (collect_spec_into_cfg, load_spec_from_cfg, validate_spec_app_settings,
  sync_spec_to_app_settings, resync_spec_tk_vars_from_app_settings,
  coerce_spec_draft_selected_gpus). Exposes SPEC_BOOL_KEYS / SPEC_STR_KEYS
  as the single source of truth for what's persisted. ConfigManager
  delegates to these at all save/load sites.
- modules/spec_tab.py (1140 lines): SpecTab class — 33 Tk vars,
  _SPEC_TYPES_* constants, full setup_tab UI builder (listbox picker,
  common controls with reset button, draft model with smart slider +
  device checkboxes + KV cache combos, ngram sections, suffix,
  ik_llama extras, vision toggle), and all 16 spec/MTP handler methods.
  Mirrors the IkLlamaTab / EnvironmentalVariablesTab pattern.

Net line deltas
---------------
- llamacpp-server-launcher.py: 5495 → 4495 (−1000)
- modules/launch.py: 1703 → 1340 (−363)
- modules/config.py: 1375 → 1193 (−182)
- Total reduction in monoliths: −1545 lines.
- New focused modules: +1940 lines (mostly class boilerplate / docstrings /
  __init__ wiring overhead for the SpecTab class).
- LaunchManager class is net SMALLER.

Test compatibility
------------------
- tests/launchers/test_spec_emission.py: redirected
  entry_module.LlamaCppLauncher._<method> → entry_module.SpecTab._<method>
  for _apply_spec_defaults_if_blank, _set_spec_draft_gpu_layers,
  _validate_spec_draft_gpu_layers_entry, _apply_mtp_parallel_default,
  _reset_spec_defaults (5 replace_all edits).
- tests/launchers/test_reasoning_and_kvu.py: redirected
  _refresh_spec_tab_state calls to SpecTab.
- _SPEC_TYPES_* references stayed unchanged (re-aliased as class attrs
  on LlamaCppLauncher for back-compat).

The launcher re-exports SpecTab's Tk vars on itself (one-time bind in
__init__ after `self.spec_tab = SpecTab(self)`) so all other code —
modules/launch.py, modules/config.py, tests with launcher-like mocks —
continues to work without imports churn. _spec_widgets / _spec_sections
are mutated in-place so the static re-export keeps a valid reference.

Tests: 1000 passing (baseline match). Ruff 0.6.9 clean. Boot smoke clean.
